requirements.txt
chevron==0.14.0
gitdb==4.0.9
GitPython==3.1.26
smmap==5.0.0
PyYAML==5.4.1
jinja2==3.1.2
jinja2-workarounds==0.1.0